20.07.06 callback 함수 추가 -OnDestroy -OnStart -OnChangePosition

+++ DirectX/Framework/Meshes/Terrain.h
@@ -54,8 +54,8 @@ public:
 class Terrain : public RenderingNode
 {
 public:
-	static Terrain* Create(float horizontal, float vertical);
-	bool Init(float horizontal, float vertical);
+	static Terrain* Create(UINT horizontal, UINT vertical);
+	bool Init(UINT horizontal, UINT vertical);
 public:
 	typedef VertexTextureNormal TerrainVertex;
